cohorte-home/requirements.txt
iPOPO==0.6.4
jsonrpclib-pelix==0.2.6
sleekxmpp==1.3.1
requests==2.11.1
Cohorte-Herald==1.0.1
Cohorte-Python==1.2.0


